requirements.txt
numpy==2.2.2
opencv-python==4.11.0.86
matplotlib==3.10.0
pytest==7.2.2
sqlalchemy==2.0.37
psycopg2==2.9.10
python-docx==1.1.2